Fix video and klv bundle package imports
DDF's upgrade of Tika to 1.24.1 introduced an updated transitive
dependency: com.sun.activation:jakarta.activation:jar:1.2.1. This caused
the javax.activation package import of the video and klv bundles to
require a minimum version of 1.2 even though the version installed in
DDF is 1.1. The bundles do not actually require version 1.2 themselves
so the version is now forced to 1.1.